🐛  [FIX] #152 Hydration mismatch 에러로 Header 기능 분리
